* strop.c (s_scm_string_capitalize_x): fix docstring quoting.
* socket.c (s_scm_inet_pton): fix docstring quoting.
(s_scm_inet_ntop): ditto.

* num2integral.i.c (INTEGRAL2NUM): cast to fix a warning.

* hashtab.c (scm_internal_hash_fold): fix argument position in
SCM_ASSERT.

* environments.c (s_scm_import_environment_set_imports_x): fix
argument position in SCM_ASSERT.

* debug.c (s_scm_make_gloc): fix SCM packing/unpacking.
(s_scm_make_iloc): ditto.

+++ libguile/num2integral.i.c
@@ -92,7 +92,7 @@ INTEGRAL2NUM (ITYPE n)
       SCM_POSFIXABLE (n)
 #endif 
       )
-    return SCM_MAKINUM (n);
+    return SCM_MAKINUM ((long) n);
 
 #ifdef SCM_BIGDIG
   return INTEGRAL2BIG (n);
